SAEM FOR LOGISTIC REGRESSION WITH MISSING DATA
Logistic regression is a reference method in supervised learning but as surprising as it may seem, there are very few solutions for performing a logistic regression and selecting variables with missing data. We suggest a stochastic approximated version of EM algorithm based on Metropolis-Hasting sampling, in order to do statistical in- ference for the logistic regression model with incomplete data. We propose a complete approach including the estimation of parameters and their variance to derive confidence interval, as well as a model se- lection procedure and how to handle missing values in testing set. The method is computationally efficient, and its good coverage and prop- erties of variable selection are demonstrated through a simulation study. We illustrate the method on a register from Paris’s hospitals on polytraumatized patients to predict the occurrence of an hem- orrhagic shock, a leading cause of early preventable death in severe trauma. The aim is to consolidate the existing Red Flag procedure, a binary alert identifying patients with high risk of severe hemorrhage.